feat: extract SchemaManager from ElasticsearchExporter.
